setup fixtures and loaders to toolkit
