+++ tests/conftest.py
@@ -1,9 +1,5 @@
-import importlib
-import sys
 import time
 from pathlib import Path
-from types import ModuleType
-from typing import Optional
 
 import numpy as np
 import pytest
@@ -12,6 +8,7 @@ from virtual_drive.core import VirtualDrive
 
 from ingeniamotion import MotionController
 from ingeniamotion.enums import SensorType
+from tests.tests_toolkit import dynamic_loader
 from tests.tests_toolkit.setups.descriptors import (
     DriveCanOpenSetup,
     DriveEcatSetup,
@@ -20,24 +17,22 @@ from tests.tests_toolkit.setups.descriptors import (
     EthercatMultiSlaveSetup,
     SetupDescriptor,
     VirtualDriveSetup,
-    descriptor_from_specifier,
 )
 from tests.tests_toolkit.setups.environment_control import (
     ManualUserEnvironmentController,
     RackServiceEnvironmentController,
     VirtualDriveEnvironmentController,
 )
-from tests.tests_toolkit.setups.rack_service_client import RackServiceClient
 from tests.tests_toolkit.setups.specifiers import (
     LocalDriveConfigSpecifier,
     MultiLocalDriveConfigSpecifier,
     MultiRackServiceConfigSpecifier,
-    RackServiceConfigSpecifier,
     SetupSpecifier,
 )
 
 pytest_plugins = [
     "tests.tests_toolkit.pytest_addoptions",
+    "tests.tests_toolkit.setup_fixtures",
 ]
 
 # Pytest runs with importlib import mode, which means that it will run the tests with the installed
@@ -47,87 +42,7 @@ pytest_plugins = [
 # be imported and ARE NOT part of the package should be specified here
 _DYNAMIC_MODULES_IMPORT = ["tests", "examples"]
 
-
-def import_module_from_local_path(
-    module_name: str, module_path: Path, add_to_sys_modules: bool = True
-) -> Optional[ModuleType]:
-    """Imports a module using a local path.
-
-    Args:
-        module_name: name that it will have in `sys.modules`.
-        module_path: module path.
-        add_to_sys_modules (bool, optional): True to add the module to `sys.modules`,
-            False otherwise. Defaults to True.
-
-    Returns:
-        imported module, None if it can not be imported.
-    """
-    # Check if it is already loaded
-    module = sys.modules.get(module_name)
-    if module is not None and hasattr(module, "__file__") and Path(module.__file__) == module_path:
-        return sys.modules[module_name]
-
-    # Load the module and add it to sys modules if required
-    if module_path.is_dir():
-        module_name = module_path.name
-        module_path /= "__init__.py"
-        if not module_path.exists():
-            return None
-    else:
-        module_name = module_path.stem
-
-    spec = importlib.util.spec_from_file_location(module_name, module_path)
-    module = importlib.util.module_from_spec(spec)
-    spec.loader.exec_module(module)
-    if add_to_sys_modules:
-        sys.modules[module_name] = module
-    return module
-
-
-def dynamic_loader(module_path: Path) -> None:
-    """Dynamically loads the modules that are not part of the package.
-
-    Should only be used if import mode is `importlib`.
-
-    Args:
-        module_path: path to the module to be loaded.
-
-    Raises:
-        RuntimeError: if the provided path is not a directory with a valid `__init__.py` file.
-    """
-    if not (module_path / "__init__.py").exists():
-        raise RuntimeError("Only directories with init path can be added in the dynamic loader")
-
-    import_module_from_local_path(module_name=module_path.name, module_path=module_path)
-
-    # Import all the submodules and files from the indicated module
-    for path in module_path.rglob("*"):
-        if "__pycache__" in path.parts:
-            continue  # Skip __pycache__ and its contents
-        elif path.is_dir():
-            if not (path / "__init__.py").exists():
-                continue  # Ignore directories without init ifle
-            sys_module_name = ".".join(path.relative_to(module_path).parts)
-        elif path.suffix == ".py":
-            if path.name == "__init__.py" or path.name.startswith("test_"):
-                continue  # Skip init file and avoid including all tests as modules
-            sys_module_name = ".".join(path.relative_to(module_path).with_suffix("").parts)
-        else:
-            continue
-
-        import_module_from_local_path(module_name=sys_module_name, module_path=path)
-
-
 test_report_key = pytest.StashKey[dict[str, pytest.CollectReport]]()
-
-
-def pytest_addoption(parser):
-    parser.addoption(
-        "--job_name",
-        action="store",
-        default="ingeniamotion - Unknown",
-        help="Name of the executing job. Will be set to rack service to have more info of the logs",
-    )
 
 
 def pytest_sessionstart(session):
@@ -141,28 +56,6 @@ def pytest_sessionstart(session):
     ingeniamotion_base_path = Path(__file__).parents[1]
     for module_name in _DYNAMIC_MODULES_IMPORT:
         dynamic_loader((ingeniamotion_base_path / module_name).resolve())
-
-
-@pytest.fixture(scope="session")
-def setup_specifier(request) -> SetupSpecifier:
-    setup_location = Path(request.config.getoption("--setup").replace(".", "/"))
-    setup_module = import_module_from_local_path(
-        module_name=setup_location.parent.name,
-        module_path=setup_location.parent.with_suffix(".py").resolve(),
-    )
-    return getattr(setup_module, setup_location.name)
-
-
-@pytest.fixture(scope="session")
-def setup_descriptor(setup_specifier, request) -> SetupSpecifier:
-    if isinstance(setup_specifier, (RackServiceConfigSpecifier, MultiRackServiceConfigSpecifier)):
-        rack_service_client = request.getfixturevalue("connect_to_rack_service")
-    else:
-        rack_service_client = None
-
-    return descriptor_from_specifier(
-        specifier=setup_specifier, rack_service_client=rack_service_client
-    )
 
 
 def connect_ethernet(mc, config, alias):
@@ -380,13 +273,6 @@ def load_configuration_after_each_module(motion_controller, setup_descriptor: Se
         mc.motion.motor_disable(servo=alias)
         if setup_descriptor.config_file is not None:
             mc.configuration.load_configuration(setup_descriptor.config_file, servo=alias)
-
-
-@pytest.fixture(scope="session")
-def connect_to_rack_service(request):
-    rack_service_client = RackServiceClient(job_name=request.config.getoption("--job_name"))
-    yield rack_service_client
-    rack_service_client.teardown()
 
 
 def __load_fw_with_protocol(mc, descriptor):
