+++ tests/tests_toolkit/setups/__init__.py
@@ -1,0 +1,14 @@
+from tests.tests_toolkit.setups.descriptors import SetupDescriptor, descriptor_from_specifier
+from tests.tests_toolkit.setups.specifiers import (
+    MultiRackServiceConfigSpecifier,
+    RackServiceConfigSpecifier,
+    SetupSpecifier,
+)
+
+__all__ = [
+    "descriptor_from_specifier",
+    "SetupDescriptor",
+    "SetupSpecifier",
+    "RackServiceConfigSpecifier",
+    "MultiRackServiceConfigSpecifier",
+]
